Move towards better config
