+++ test/test_config.py
@@ -1,7 +1,7 @@
 #  test to_dict and from_dict
 
 
-from gmft._dataclasses import with_config
+from gmft.config._dataclasses import with_config
 from gmft.formatters.tatr import TATRFormatConfig
